refactor: 現在位置移動系のリファクタ
